v2

+++ lib/batch/batch.go
@@ -19,7 +19,7 @@ func getBatch(n int64, pool int64) (res []user) {
 	var wg sync.WaitGroup
 
 	sem := make(chan struct{}, pool)
-	for i := 0; i < n; i++ {
+	for i := int64(0); i < n; i++ {
 		wg.Add(1)
 		go func(i int64) {
 			sem <- struct{}{}
